Add files via upload
The purpose of this Zomato rating SQL project is to analyse the data of restaurants with important attributes like restaurant name, rating, food, price average, delivery time, security system and location Step a the first is to create a SQL table to store the data, then populate it with the given information. Subsequent survey data analyses include analysing the distribution of ratings, identifying restaurants with the highest ranking, and analysing the popularity of various food items to understand safety measures species status and examine the safety measures adopted by restaurants. Place-based methods shed light on differences in ratings across sectors. In addition, the project examines in detail the relationship between delivery time and volume, and the relationship between delivery price and delivery quantity The results are not only available through SQL queries they are displayed but can also be used for data visualization of external instruments. Eventually, the project can extend the conclusions, insights, and recommendations possible in the dynamic number of restaurants and customer preferences to creating a basic recommendation system this Zomato rating SQL project, which its purpose is to find out the name of restaurant, the number of servings, Insights should be obtained from restaurant data through important attributes such as food, average price, delivery time, security measures and location on the addition of the. The first step is to create an SQL table to store the data, followed by populating it with the given information. Subsequent research data analyzes include analysis of the distribution of serving sizes, identification of the most popular restaurants, and analysis of the popularity of various foods Safety measures.